backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "faiss-cpu>=1.9.0",
    "fastapi>=0.115.0",
    "groq>=0.12.0",
    "httpx>=0.27.0",
    "numpy>=1.26.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "python-dotenv>=1.0.0",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=3.3.0",
    "slowapi>=0.1.9",
    "uvicorn[standard]>=0.32.0",
]

[tool.pytest.ini_options]
asyncio_mode = "strict"

[tool.ruff]
line-length = 100
